+++ includes/class-stancer-gateway.php
@@ -98,6 +98,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 		add_filter( 'woocommerce_order_button_html', [ $this, 'place_order_button_html' ] );
 
 		// Update settings.
+		// @phpstan-ignore-next-line we won't use the return value here but it still can be usefull.
 		add_action( 'woocommerce_update_options_payment_gateways_' . $this->id, [ $this, 'process_admin_options' ] );
 
 		// Check the order status and redirect us if the payment is already completed.
@@ -119,7 +120,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * @param WC_Order $order Order.
 	 * @param string|null $card_id Card identifier.
 	 *
-	 * @return array
+	 * @return CheckoutResponse
 	 */
 	public function create_api_payment( $order, $card_id = null ) {
 		$redirect = $order->get_checkout_payment_url( true );
@@ -196,6 +197,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * Display notice when payment failed.
 	 *
 	 * @since 1.0.0
+	 *
+	 * @return void
 	 */
 	public function display_notice() {
 		$notice = WC()->session->get( 'stancer_error_payment' );
@@ -211,7 +214,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @since 1.1.0
 	 *
-	 * @return self
+	 * @return static
 	 */
 	public function dynamic_title() {
 		global $wpdb;
@@ -225,7 +228,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 			$this->title = 'Stancer';
 		}
 
-		$table = fn( string $name ) => '`' . $wpdb->prefix . 'wc_stancer_' . $name . '`';
+		$table = fn ( string $name ) => '`' . $wpdb->prefix . 'wc_stancer_' . $name . '`';
 
 		// phpcs:disable WordPress.DB.PreparedSQL.NotPrepared
 
@@ -277,7 +280,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 
 			$this->title = vsprintf(
 				// translators: $1 Card brand. $2 Last 4. $3 Expiration month. $4 Expiration year.
-				__( '%1$s finishing with %2$s', 'stancer' ),
+				__( '%1$s ending with %2$s', 'stancer' ),
 				[
 					$result['brand_name'],
 					$result['last4'],
@@ -286,7 +289,6 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 				],
 			);
 		}
-
 		return $this;
 	}
 
@@ -336,7 +338,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * @since 1.1.0
 	 *
 	 * @param string $key Field key.
-	 * @param array $data Field data.
+	 * @param FormField $data Field data.
+	 *
 	 * @return string
 	 */
 	public function generate_payment_option_logo_html( $key, $data ) {
@@ -434,7 +437,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * Get the payment Data to be send to our frontEnd.
 	 *
 	 * @param string $page_type the page type chosen by the user.
-	 * @return array
+	 *
+	 * @return GetPaymentData
 	 */
 	public function get_payment_data( string $page_type ): array {
 		$data = [
@@ -464,10 +468,10 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @return array
+	 * @return KeyConfigurationData
 	 */
 	public function get_configurations() {
-		// translators: %s: Key prefixes (aka sprod, pprod, stest or ptest).
+		// translators: "%s": Key prefixes (aka sprod, pprod, stest or ptest).
 		$desc = __( 'Starts with "%s"', 'stancer' );
 
 		return [
@@ -505,7 +509,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * @since 1.2.0 Add admin scripts.
 	 * @since 1.2.5 Admin scripts moved to WC_Stancer_Gateway::form_fields_scripts.
 	 *
-	 * @return self
+	 * @return static
 	 */
 	public function init_form_fields() {
 
@@ -652,17 +656,11 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 			'label' => __( 'Page type', 'stancer' ),
 			'title' => __( 'Page type', 'stancer' ),
 			'options' => [
-				'iframe' => __( 'Popup (Deprecated)', 'stancer' ),
 				'pip' => __( 'Inside the page', 'stancer' ),
 				'redirect' => __( 'Redirect to an external page', 'stancer' ),
 			],
 			'type' => 'select',
 		];
-
-		if ( array_key_exists( 'page_type', $this->settings ) && 'iframe' === $this->settings['page_type'] ) {
-			$description = __( 'Popup option is deprecated and will be removed in a future update.', 'stancer' );
-			$inputs['page_type'] = array_merge( $inputs['page_type'], [ 'description' => $description ] );
-		}
 
 		$desc_auth_limit = __(
 			'Minimum amount to trigger an authenticated payment (3DS, Verified by Visa, Mastercard Secure Code...).',
@@ -731,6 +729,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @since 1.0.0
 	 * @since 1.1.0 Add logos and description.
+	 *
+	 * @return void
 	 */
 	public function payment_fields() {
 		$logo = $this->settings['payment_option_logo'] ?? 'no-logo';
@@ -758,8 +758,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 		$page_type = $payment_data['page_type'];
 		$data = $payment_data['data'];
 
-		$script_path = fn( string $name ) => plugin_dir_url( STANCER_FILE ) . 'public/js/' . $name . '.min.js';
-		$style_path = fn( string $name ) => plugin_dir_url( STANCER_FILE ) . 'public/css/' . $name . '.min.css';
+		$script_path = fn ( string $name ) => plugin_dir_url( STANCER_FILE ) . 'public/js/' . $name . '.min.js';
+		$style_path = fn ( string $name ) => plugin_dir_url( STANCER_FILE ) . 'public/css/' . $name . '.min.css';
 
 		$add_script = function ( string $script, bool $localize = true, array $dependancy = [] ) use ( $data, $script_path ) {
 			$name = 'stancer-' . $script;
@@ -779,10 +779,6 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 		};
 		echo esc_html( $payment_data['message'] );
 		switch ( $page_type ) {
-			case 'iframe':
-				$add_script( 'popup' );
-				break;
-
 			case 'pip':
 				$dependancy[] = $add_script( 'api', false );
 				$dependancy[] = $add_script( 'change_payment_method', false );
@@ -827,7 +823,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @param int $order_id Order ID.
 	 *
-	 * @return array
+	 * @return CheckoutResponse
 	 */
 	public function process_payment( $order_id ) {
 		$order = wc_get_order( $order_id );
@@ -841,6 +837,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * @since 1.0.0
 	 *
 	 * @param int $order_id Order ID.
+	 *
+	 * @return void|never
 	 */
 	public function receipt_page( $order_id ) {
 
@@ -918,7 +916,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 
 				$order->add_order_note(
 					sprintf(
-						// translators: %s: Stancer payment identifier.
+						// translators: "%s": Stancer payment identifier.
 						__( 'Payment was completed via Stancer (Transaction ID: %s)', 'stancer' ),
 						$api_payment->getId()
 					)
@@ -939,7 +937,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 * Redirect Processing order to order_Checkout .
 	 * We can have multiple redirection but if they scope badly we have a processing Error.
 	 *
-	 * @return void
+	 * @return void|never
 	 */
 	public function redirect_incorrect_call() {
 		$finished_status = [
@@ -967,7 +965,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @param string $value The value of the description.
 	 * @param string $default_message The default message, if description is not valid.
-	 * @return void
+	 *
+	 * @return void|string
 	 */
 	protected function validate_description( $value, $default_message ) {
 		if ( ! $value ) {
@@ -998,6 +997,7 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @param string $key the key of the settings.
 	 * @param string $value the value of payment description.
+	 *
 	 * @return string The value of payment description.
 	 */
 	public function validate_payment_description_field( $key, $value ) {
@@ -1014,7 +1014,8 @@ class WC_Stancer_Gateway extends WC_Payment_Gateway {
 	 *
 	 * @param string $key the key of the settings.
 	 * @param string $value the value of payment description.
-	 * @return string The value of payment description.
+	 *
+	 * @return void|string The value of payment description.
 	 */
 	public function validate_subscription_renewal_description_field( $key, $value ) {
 		if ( $this->subscriptions_enabled() ) {