+++ subscription/change-payment-method.php
@@ -30,7 +30,7 @@ nocache_headers();
  * @param Stancer\Payment $payment Payment used.
  */
 function create_card_info( Stancer\Payment $payment ): string {
-	// translators: $1 Card brand. $2 Last 4. $3 Expiration month. $4 Expiration year.
+	// translators: "%1$s": Card brand. "%2$s": Last 4.
 	$trad = __( '%1$s finishing with %2$s', 'stancer' );
 	$card = $payment->card;
 
@@ -40,7 +40,9 @@ function create_card_info( Stancer\Payment $payment ): string {
 /**
  * Helper to send the response.
  *
- * @param array $data Response data.
+ * @param array<mixed> $data Response data.
+ *
+ * @return void
  */
 function send_response( array $data ) {
 	echo wp_json_encode( $data );
@@ -164,7 +166,7 @@ try {
 
 			break;
 		default:
-			throw new Stancer\Exceptions\Exception( __( 'Incorrect action method', 'stancer' ) );
+			throw new Stancer\Exceptions\Exception( __( 'Incorrect action', 'stancer' ) );
 	}
 } catch ( Stancer\Exceptions\Exception $exception ) {
 	$response['reason'] = $exception->getMessage();