+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-08-01
+
+### Added
+- Support for Wordpress version 6.7 and 6.8 (CMS-374)
+- Better versionning for the devcontainer
+- PHPStan static analysis (CMS-96)
+
+### Fixed
+- Translation harmonization between CMS (CMS-273)
+
+### Removed
+- Support for Wordpress version 6.1 to 6.4 (CMS-374)
+- Remove the deprecated popup option (CMS-79)
+
+
 ## [1.3.1] - 2025-07-28
 
 ### Fixed
@@ -14,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.0] - 2024-09-26
 
 ### Added
-- Support of woocommerce checkout blocks. (CMS-68)
+- Support of woocommerce checkout blocks (CMS-68)
 
 ### Changed
 - Change the cb logo to the new logo (CMS-173)